fix: resolve found issues
